another fix

+++ tutorial/createdefinition.sh
@@ -1,4 +1,4 @@
-#! /bin/sh
+#!/usr/bin/env bash
 
 ###############################################################
 # Copyright 2010 – 2024 Rocket Software, Inc. or its affiliates.
@@ -18,31 +18,28 @@
 ###############################################################
 
 SCRIPTDIR="$(dirname "$(readlink -f "$0")")"
-if [ $# = 0 ]; then
+if [[ $# -eq 0 ]]; then
     FILEPATH="$SCRIPTDIR"
 else
     FILEPATH=$1
 fi
 
-if [ ! -d "$FILEPATH" ]; then
-    echo "WARNING: Path does not exist."
+if [[ ! -d "$FILEPATH" ]]; then
+    echo "WARNING: Path does not exist." >&2
 fi
 WRITEXML="y"
-if [ -f "$SCRIPTDIR/BANKDEMO.xml" ]; then
-    if [ -f "$SCRIPTDIR/BANKDEMO.xml.bak" ]; then
+if [[ -f "$SCRIPTDIR/BANKDEMO.xml" && -f "$SCRIPTDIR/BANKDEMO.xml.bak" ]]; then
         echo "$SCRIPTDIR/BANKDEMO.xml.bak already exists. Do you want to override? (y/n) "
         read WRITEXML
-        if [ "$WRITEXML" = "y" -o "$WRITEXML" = "Y" ]; then
+        if [[ "$WRITEXML" == "y" || "$WRITEXML" == "Y" ]]; then
             cp "$SCRIPTDIR/BANKDEMO.xml" "$SCRIPTDIR/BANKDEMO.xml.bak"
         fi
-    fi
-
 fi
-if [ "$WRITEXML" = "y" -o "$WRITEXML" = "Y" ]; then
+if [[ "$WRITEXML" == "y" || "$WRITEXML" == "Y" ]]; then
     cp "$SCRIPTDIR/BANKDEMO.template" "$SCRIPTDIR/BANKDEMO.xml"
     sed -i -e "s!__IMPORT_FILE_DIR__!$FILEPATH!" "$SCRIPTDIR/BANKDEMO.xml" 
     sed -i -e "s!PS=;!PS=:!" "$SCRIPTDIR/BANKDEMO.xml"
     sed -i -e "s!OS=Windows!OS=Linux!" "$SCRIPTDIR/BANKDEMO.xml"
 else
-    echo "Error: Enterprise Server definition was not created."
+    echo "Error: Enterprise Server definition was not created." >&2
 fi
